Refactor Player.cs, update SpawnManager.cs, change primary Scene
Player.cs was cleaned up and refactored for readability and performance. Unnecessary spaces and lines were removed, and the enemy component is now rechecked every two seconds instead of five for a more responsive gameplay experience.

In SpawnManager.cs, a limitation was added to prevent the system from endlessly spawning enemies. The new maxEnemiesToSpawn variable provides an easy way to control the maximum number of enemies on the field at one time.

Finally, the primary scene was switched from SampleScene.unity to Town.unity in EditorBuildSettings.asset. Now, when the game is run, it will always start from the town scene, which is more appropriate as an initial game stage.